+++ internal/logic/install.go
@@ -96,7 +96,8 @@ debug: false
 # Reward >= v0.2.34 uses the internally built "docker.io/rewardenv/sshd"
 #reward_tunnel_image: "docker.io/rewardenv/sshd"
 
-# Override default traefik ports
+# Override default listen address and ports for traefik
+#reward_traefik_listen: "0.0.0.0"
 #reward_traefik_http_port: 80
 #reward_traefik_https_port: 443
 
@@ -108,14 +109,27 @@ reward_traefik_bind_additional_http_ports: []
 # reward_traefik_bind_additional_https_ports: [8443,9443]
 reward_traefik_bind_additional_https_ports: []
 
+# By default, Reward redirects all HTTP traffic to HTTPS. If you want to disable this behaviour, you can
+# uncomment the following line.
+#reward_traefik_allow_http: false
+
 # By default Reward makes it possible to resolve the environment's domain to the nginx container's IP address
 # inside the docker network. To disable this behaviour you can uncomment the following line.
 #reward_resolve_domain_to_traefik: false
+
+# Override default dnsmasq listen address and ports
+#reward_dnsmasq_listen: "0.0.0.0"
+#reward_dnsmasq_tcp_port: "53"
+#reward_dnsmasq_udp_port: "53"
 
 # By default, only the UDP port 53 is exposed from the dnsmasq container. Sometimes it doesn't seem to be enough, and
 # the TCP port 53 has to be exposed as well. To do so enable the "reward_dnsmasq_bind_tcp" variable.
 #reward_dnsmasq_bind_tcp: true
 #reward_dnsmasq_bind_udp: true
+
+# Override default tunnel listen address and ports
+#reward_tunnel_listen: "0.0.0.0"
+#reward_tunnel_port: "2222"
 
 # By default Reward is not allowed to run commands as root.
 # To disable this check you can uncomment the following line.
@@ -336,9 +350,11 @@ func (c *installer) checkInstalled() {
 	if !c.installCaCertFlag() && !c.installDNSFlag() && !c.installSSHKeyFlag() && !c.installSSHConfigFlag() {
 		if util.FileExists(c.InstallMarkerFilePath()) {
 			if !util.AskForConfirmation(
-				fmt.Sprintf("%s is already installed. Would you like to reinstall?",
+				fmt.Sprintf(
+					"%s is already installed. Would you like to reinstall?",
 					cases.Title(language.English).String(c.AppName()),
-				)) {
+				),
+			) {
 				log.Println("...installation aborted.")
 				os.Exit(0)
 			}
@@ -732,8 +748,12 @@ Host tunnel.%[2]s.test
 
 			log.Debugf("Searching for configuration regex in ssh config file: %s...", sshConfigFile)
 
-			matches := regexp.MustCompile(fmt.Sprintf("## %s START ##",
-				strings.ToUpper(c.AppName()))).FindStringSubmatch(string(content))
+			matches := regexp.MustCompile(
+				fmt.Sprintf(
+					"## %s START ##",
+					strings.ToUpper(c.AppName()),
+				),
+			).FindStringSubmatch(string(content))
 
 			log.Debugf("...regex match: %+v", matches)